rename package pkg/metainfo/kvmetainfo to uplink/metainfo/kvmetainfo
