+++ lib/uplink/uplink.go
@@ -11,10 +11,10 @@ import (
 
 	"storj.io/storj/internal/memory"
 	"storj.io/storj/pkg/identity"
-	"storj.io/storj/pkg/metainfo/kvmetainfo"
 	"storj.io/storj/pkg/peertls/tlsopts"
 	"storj.io/storj/pkg/transport"
 	"storj.io/storj/uplink/metainfo"
+	"storj.io/storj/uplink/metainfo/kvmetainfo"
 )
 
 const defaultUplinkDialTimeout = 20 * time.Second
